Split components and close dialog on upload
